key 값에 따른 정책 설정(write,get,edit,delete),write 100개 초과 제한

+++ src/main/java/com/crudtest/service/PostService.java
@@ -2,6 +2,8 @@ package com.crudtest.service;
 
 import com.crudtest.domain.Post;
 import com.crudtest.domain.PostEditor;
+import com.crudtest.exception.LimitPost;
+import com.crudtest.exception.PostNotAuthority;
 import com.crudtest.exception.PostNotFound;
 import com.crudtest.repository.PostRepository;
 import com.crudtest.request.PostCreate;
@@ -23,8 +25,11 @@ public class PostService {
     private final PostRepository postRepository;
 
     public Long write(PostCreate postCreate,String key){
+        List<Post> posts = postRepository.findAllByKey(key);
+        if(posts.size() == 100) {
+            throw new LimitPost();
+        }
         //postCreate => Entity
-
         Post post = Post.builder()
                 .title(postCreate.getTitle())
                 .content(postCreate.getContent())
@@ -36,9 +41,13 @@ public class PostService {
         return post.getId();
     }
 
-    public PostResponse get(Long id){
+    public PostResponse get(Long id,String key){
         Post post = postRepository.findById(id)
                 .orElseThrow(PostNotFound::new);
+
+        if(!post.getKey().equals(key)) {
+            throw new PostNotAuthority("Id","확인할 권한이 없습니다.");
+        }
 
         return PostResponse.builder()
                 .id(post.getId())
@@ -50,17 +59,20 @@ public class PostService {
     //글이 너무 많은 경우 -> 비용이 너무 많이 든다.
     //글이 100,000,000 개인 경우 -> DB가 뻗을 수 있다.
     //DB -> 애플리케이션 서버로 전달되는 시간 ,트래픽 비용 발생
-    public List<PostResponse> getList(PostSearch postSearch){
-        return postRepository.getList(postSearch).stream()
+    public List<PostResponse> getList(PostSearch postSearch,String key){
+        return postRepository.getList(postSearch,key).stream()
                 .map(PostResponse::new)
                 .collect(Collectors.toList());
     }
 
     @Transactional
-    public void edit(Long id, PostEdit postEdit){
+    public void edit(Long id, PostEdit postEdit,String key){
         Post post = postRepository.findById(id)
                 .orElseThrow(PostNotFound::new);
 
+        if(!post.getKey().equals(key)) {
+            throw new PostNotAuthority("Id","수정할 권한이 없습니다.");
+        }
         PostEditor.PostEditorBuilder editorBuilder = post.toEditor();
 
         PostEditor postEditor = editorBuilder.title(postEdit.getTitle())
@@ -70,9 +82,13 @@ public class PostService {
         post.edit(postEditor);
     }
 
-    public void delete(Long id) {
+    public void delete(Long id,String key) {
         Post post = postRepository.findById(id)
                 .orElseThrow(PostNotFound::new);
+
+        if(!post.getKey().equals(key)) {
+            throw new PostNotAuthority("Id","삭제할 권한이 없습니다.");
+        }
 
         postRepository.delete(post);
     }
